optimization-part-1: Refactor current architectire, files structure and classes name
