--- .github/ISSUE_TEMPLATE.md
@@ -1,2 +1,0 @@
-<!-- Love flyimg? Please consider supporting our collective:
-👉  https://opencollective.com/flyimg/donate -->
