+++ tests/Core/Service/ImageProcessorTest.php
@@ -12,7 +12,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessPNG()
     {
-        $image = $this->coreManager->processImage(parent::OPTION_URL.',o_png', parent::PNG_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::OPTION_URL.',o_png', parent::PNG_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
         $this->assertEquals(Image::PNG_MIME_TYPE, $this->getFileMemeType($image->getNewFilePath()));
@@ -22,7 +22,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessWebpFromPng()
     {
-        $image = $this->coreManager->processImage(parent::OPTION_URL.',o_webp', parent::PNG_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::OPTION_URL.',o_webp', parent::PNG_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
         $this->assertEquals(Image::WEBP_MIME_TYPE, $this->getFileMemeType($image->getNewFilePath()));
@@ -32,7 +32,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessJpgFromPng()
     {
-        $image = $this->coreManager->processImage(parent::OPTION_URL.',o_jpg', parent::PNG_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::OPTION_URL.',o_jpg', parent::PNG_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
         $this->assertEquals(Image::JPEG_MIME_TYPE, $this->getFileMemeType($image->getNewFilePath()));
@@ -42,7 +42,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessGifFromPng()
     {
-        $image = $this->coreManager->processImage(parent::OPTION_URL.',o_gif', parent::PNG_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::OPTION_URL.',o_gif', parent::PNG_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
         $this->assertEquals(Image::GIF_MIME_TYPE, $this->getFileMemeType($image->getNewFilePath()));
@@ -52,7 +52,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessJpg()
     {
-        $image = $this->coreManager->processImage(parent::OPTION_URL, parent::JPG_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::OPTION_URL, parent::JPG_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
     }
@@ -61,7 +61,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessGif()
     {
-        $image = $this->coreManager->processImage(parent::GIF_OPTION_URL, parent::GIF_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::GIF_OPTION_URL, parent::GIF_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
         $this->assertEquals(Image::GIF_MIME_TYPE, $this->getFileMemeType($image->getNewFilePath()));
@@ -71,7 +71,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessPngFromGif()
     {
-        $image = $this->coreManager->processImage(parent::GIF_OPTION_URL.',o_png', parent::GIF_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::GIF_OPTION_URL.',o_png', parent::GIF_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
         $this->assertEquals(Image::PNG_MIME_TYPE, $this->getFileMemeType($image->getNewFilePath()));
@@ -81,7 +81,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessJpgFromGif()
     {
-        $image = $this->coreManager->processImage(parent::GIF_OPTION_URL.',o_jpg', parent::GIF_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::GIF_OPTION_URL.',o_jpg', parent::GIF_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
         $this->assertEquals(Image::JPEG_MIME_TYPE, $this->getFileMemeType($image->getNewFilePath()));
@@ -91,7 +91,7 @@ class ImageProcessorTest extends BaseTest
      */
     public function testProcessWebpFromGif()
     {
-        $image = $this->coreManager->processImage(parent::GIF_OPTION_URL.',o_webp', parent::GIF_TEST_IMAGE);
+        $image = $this->ImageHandler->processImage(parent::GIF_OPTION_URL.',o_webp', parent::GIF_TEST_IMAGE);
         $this->generatedImage[] = $image;
         $this->assertFileExists($image->getNewFilePath());
         $this->assertEquals(Image::WEBP_MIME_TYPE, $this->getFileMemeType($image->getNewFilePath()));
